requirements.txt
aniso8601==8.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
ffmpeg-python==0.2.0
Flask==1.1.1
Flask-RESTful==0.3.8
future==0.18.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.4.0
oauthlib==3.1.0
Pillow==7.0.0
PySocks==1.7.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
tweepy==3.8.0
urllib3==1.25.8
webargs==5.5.3
Werkzeug==1.0.0
